Delete uuid utils file, moved functions to utils.rs
