+++ src/utils/mod.rs
@@ -1,1 +1,2 @@
-pub mod uuid;
+pub mod fetch;
+pub mod utils;
